Moved all logic to be handled within SeedingService HTHL related SeedDatabase() in HomeController Removed SeedingController
Cleaned it up, added comments when needed.
