+++ Controllers/HomeController.cs
@@ -2,6 +2,7 @@ using System.Diagnostics;
 using Microsoft.AspNetCore.Mvc;
 using Garage.Models;
 using Garage.Services;
+using Garage.Data;
 
 namespace Garage.Controllers;
 
@@ -9,11 +10,13 @@ public class HomeController : Controller
 {
     private readonly IGarageService _garageService;
     private readonly ILogger<HomeController> _logger;
+    private readonly ISeedingService _seedingService;
 
-    public HomeController(ILogger<HomeController> logger, IGarageService garageService)
+    public HomeController(ILogger<HomeController> logger, IGarageService garageService, ISeedingService seedingService)
     {
         _logger = logger;
         _garageService = garageService;
+        _seedingService = seedingService;
     }
 
     public async Task<IActionResult> Index()
@@ -32,4 +35,17 @@ public class HomeController : Controller
     {
         return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
     }
+
+    [HttpPost]
+    public async Task<IActionResult> SeedDatabase()
+    {
+        if (!await _seedingService.IsDatabaseSeededAsync())
+        {
+            await _seedingService.SeedDatabaseAsync();
+            return RedirectToAction(nameof(Index));
+        }
+
+        return RedirectToAction(nameof(Index));
+    }
+
 }
