+++ Data/GarageDbContext.cs
@@ -24,3 +24,4 @@ public class GarageDbContext : DbContext
         options.UseSqlServer(Configuration.GetConnectionString("DefaultConnection"));
     }
 }
+
