+++ Views/Home/Index.cshtml
@@ -14,7 +14,7 @@
     }
 </ul>
 <div class="custom-container3 d-flex justify-content-between align-items-start">
-    <form asp-controller="Seed" asp-action="SeedDatabase" method="post">
+    <form asp-controller="Home" asp-action="SeedDatabase" method="post">
         <button type="submit">Seed Vehicles</button>
     </form>
 </div>
